+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.7.0] - 2020-03-11
+
+### Added
+- Add theme customizer
+- Add optionizer (for setting up options)
+- Add menus setup
+- Add metaboxes setup
+- Add post type templates setup
+- Add scripts (JavaScript) setup
+- Add sidebars setup
+- Add styles setup
+- Add thumbnails setup
+- Add translations setup
+
 ## [0.6.2] - 2020-02-08
 
 ### Changed